Separada vista de la VistaModelo
