+++ app/build.gradle
@@ -41,4 +41,10 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+
+    // para instanciar ViewModel en la Activity
+    implementation "androidx.fragment:fragment-ktx:1.5.4"
+    // para los observables
+    implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.5.1'
+
 }
